Add getActiveDemon to summoner class
Coming back to this after a bit of a break, I realise that all of the classes make inherent sense apart from the battle class, which is incredibly confusing and totally unmaintainable. I need to start again from the ground up with the battle, and part of that is going to be abstracting as many methods as I can from the battle mode into other classes where they make more sense to be in. I also need to consider the best form for the battle data structure, and whether it really makes sense to encapsulate the whole thing into one class, or whether multiple classes would be more appropriate.